+++ BankAccount/src/main/java/io/github/LuizYokoyama/BankAccount/dto/EntryDto.java
@@ -3,6 +3,7 @@ package io.github.LuizYokoyama.BankAccount.dto;
 import io.github.LuizYokoyama.BankAccount.entity.OperationType;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -20,7 +21,7 @@ public class EntryDto {
 
     private OperationType operationType;
 
-    private float value;
+    private BigDecimal value;
 
     private LocalDateTime entryDateTime;
